+++ resources/unmanaged_device.rb
@@ -1,4 +1,4 @@
-# (c) Copyright 2016 Hewlett Packard Enterprise Development LP
+# (c) Copyright 2016-2017 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,19 +13,14 @@ OneviewCookbook::ResourceBaseProperties.load(self)
 
 default_action :add
 
-action_class do
-  include OneviewCookbook::Helper
-  include OneviewCookbook::ResourceBase
-end
-
 action :add do
-  add_or_edit
+  OneviewCookbook::Helper.do_resource_action(self, :UnmanagedDevice, :add_or_edit)
 end
 
 action :add_if_missing do
-  add_if_missing
+  OneviewCookbook::Helper.do_resource_action(self, :UnmanagedDevice, :add_if_missing)
 end
 
 action :remove do
-  remove
+  OneviewCookbook::Helper.do_resource_action(self, :UnmanagedDevice, :remove)
 end