+++ CHANGELOG.md
@@ -7,6 +7,11 @@
   - datacenter support for API300
   - fabric new resource support for API300 Synergy
   - san_manager support for API300
+  - rack support for API300
+  - managed_san support for API300
+  - server_hardware_type resource for API300
+  - power_device support for API300
+  - unmanaged device support for API300
   - logical_interconnect support for API300
 
 ### 1.1.0